Uploaded all the Project Euler solutions to date
Please note: Euler 3 was completely copied from a youtube video which is specified within the file. Not my own work. I was at a point where I did not have enough python syntax learned to complete the problem. Copying the code and fooling around with it allowed me to make huge leaps and contributed greatly to my learning.